+++ src/test/groovy/life/qbic/infrastructure/event/serialization/EventSerializerFactorySpec.groovy
@@ -1,4 +1,4 @@
-package life.qbic.infrastructure.serialization.event
+package life.qbic.infrastructure.event.serialization
 
 import life.qbic.domain.sample.SampleCode
 import life.qbic.domain.sample.events.*
